db: read the minimum to determine which shard a sstable belongs to
Currently, we only determine if a sstable belongs to current shard
after loading some of its components into memory. It turns out that
the filter size may be several megabytes long for big sstables.
So this patch is intended to optimize the loading process by reading
the minimum from disk to determine whether a sstable belongs to the
current shard or not.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>