db: introduce function to optimize sstable loading process
Currently, we only determine if a sstable belongs to current shard
after loading some of its components into memory. For example,
filter may be considerably big and its content is irrelevant to
decide if a sstable should be included to a given shard.
So this patch introduces a function to optimize the loading process
by reading the minimum from disk to determine whether a sstable
belongs to the current shard or not.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>