sstables: introduce sstable::rewrite
SSTable rewrite was implemented very similar to SSTable compaction.
There is a read fiber that reads mutations from the input sstable
and pass them through the pipe to the write fiber.
Mutation is discarded if it doesn't belong to current shard or if
its token isn't contained in any of the token ranges that belong
to the underlying node. rewrite is the basis of cleanup.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>