db: introduce column_family::perform_cleanup_on_sstables
It's intended to clean up each sstable of a given column family.
Cleanup is about discarding any keys that are not longer relevant
to a sstable, which may result in disk space saving.
This feature is frequently used after a node was added to the
cluster because existing nodes lose part of its token range.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>