compaction: filter out compacting sstables
That's needed to support more than one type of compaction on
a given column family. For example, cleanup and compaction can
be running in parallel for column family X, and we must make
sure that they don't use the same sstables.
This change is about filtering out sstables that are being
compacting, in which the information in stored in the sstable
itself (after this patch).

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>