sstables: fix statistics rewrite
It's not working because it tries to overwrite existing statistics
file with exclusive flag.
It's fixed by writing new statistics into temporary file and
renaming it into place.
We still need to adjust boot code to deal with scenario where scylla
is killed in between, in which a temporary file is left over.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>