compaction: retry compaction until strategy returns an empty job
We should retry compaction until compaction strategy says there's
nothing more to compact. At the moment, we only retry compaction
if a new sstable was written. Basically, one compaction for each
memtable flush. But that information is lost after node restart,
and consequently, column family may be left with lots of sstables
that can be compacted.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>