sstables: move sstable write functionality into a new class
Before these changes, sstable read and write functionality were living in the same class,
and that's prone to errors because an immutable sstable could be incorrectly written to.
Also, we had to make sure that a sstable just written is loaded with necessary data for
subsequent reads on it to succeed. Something like the following:
    sstable sst = sstable(...);
    sst.write_components(memtable).get();
    sst.open_data().get();

With these changes, sstable_writer will be used to write a new sstable, and when the
new sstable is sealed, a readable sstable object can be obtained from the sstable_writer
itself. For example:
    sstable_writer writer = sstable_writer(...);
    write.write_components(memtable).get();
    sstable sst = sstable_writer::get_sstable_from_writer(std::move(writer)).get0();

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>