compaction: reduce bloom filter overhead
For each compacted partition in compaction, bloom filter of non
compacting sstable is consulted. The overhead is big as shown by
the flame graph shared in the issue #1322.

To reduce this overhead, sstable_set::select() may be used to
only consult bloom filter of sstables that really matter.
If leveled strategy used, it means that we will only consult bloom
filter of sstables that overlap with the ones being compacted.
I think that this should be the default behavior, and that can be
achieved by unconditionally building a partitioned_sstable_set in
build_sstable_set().

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>