db: disable clustering optimization for leveled strategy
Leveled strategy will not benefit considerably from this strategy
because there's only a few sstables that will contain a given
partition key, which means that a clustering key that belongs to
a specific partition key can only be in a few sstables as well.
For other strategies, it will be enabled. Date tiered strategy is
the one that will benefit the most from this optimization.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>