db: enable clustering optimization only on date tiered strategy
Leveled strategy will not benefit considerably from this strategy
because there's only a few sstables that will contain a given
partition key, which means that a clustering key that belongs to
a specific partition key can only be in a few sstables as well.
Date tiered strategy is the one that will actually benefit the
most from this optimization. Size tiered may benefit from it too
if clustering key isn't overwritten.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>