streaming_histogram: do not limit it to be used by sstables
streaming histogram will later be placed in /utils, so we want
to use std::map<> instead of disk_hash<>. For that we need to
implement serialization/deserialization functions for it.
I switched the algorithm from the hash table of disk_hash<>
(std::unordered_map) to a heap (std::map) because update()
relies on keys being ordered.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>