streaming_histogram: fix update
This bug was introduced when converting java code. Return value
of map::erase() was used as if it were the value of the removed
entry, but it's actually the number of removed entries.

update() also relies on ordered keys, so map is used instead.

Existing sstables which map size is greater than 100 have an
incorrect histogram. Coming patch will detect bad histograms
and discard them.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>