db: avoid excessive memory usage during resharding
After resharding, all sstables may be owned by all shards, which
means that file descriptors and memory usage for metadata will
increase by a factor equal to number of shards. That can easily
lead to OOM or severe downtime.

SSTable components are immutable, so they can be stored in one
shard and shared with others that need it. (generation % smp
count) is used to decide which shard will store because we want
to distribute sstables evenly. A manager is introduced to
orchestrate this whole thing. It's similar to deletion manager
in nature.

We rely on the fact that shard 0 is populated first to guarantee
that all shards will be able to find the shared components in
the manager.

Fixes #1951.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>