db: fix quadratic behavior when checking for repeated generation
To detect an attempt to load more than one sstable with the same
generation for a column family, we were iterating through all
sstable set for each sstable loaded. That results in quadratic
behavior in the number of sstables loaded.

It can be fixed by checking for repeated generation in the end of
column family population (that's fine in my understanding because
this case is very rare and this change only results in delaying the
detection) and also using a better data structure for it.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>