loader: don't rely on name of sstable to determine its column family
When opening a sstable during loading phase, we currently rely on
component keyspace and column family embedded in sstable name to
find its column family.

If column family A stores sstable of column family B by mistake,
it may happen that population of column family A will lead to
column family B loading a sstable that is stored by A.

To fix it, let's propagate the name of keyspace and column family
we are populating, so we'll always use the correct column family.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>