db: fix quadratic behavior when checking for repeated generation
To detect an attempt to load more than one sstable with the same
generation for a column family, we were iterating through all
sstable set for each sstable loaded. That results in quadratic
behavior in the number of sstables loaded.

It's fixed by using a better data structure. The check is also
moved to sstable set.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>