database: revamp sstable resharding
From now on, new resharding algorithm will be used for shared sstables.
Its benefit is that we'll no longer read the same shared sstable in
all shards that own it. Therefore, this new approach saves bandwidth.

As before, only one reshard will proceed at a time per shard.
Resharding of a column family is paralellized in all shards.

By the time being, we still only reshard one sstable at a time.
In the future, compaction strategy will help on the decision of which
sstables to reshard.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>