compaction: make leveled strategy resilient to parallel operations
This doesn't mean that compaction manager is allowing parallel
compaction for leveled strategy (at least not yet) but that
it will be made resilient against a cleanup or resharding in
parallel, which can result in overlapping in level > 0.
For example, if a sstable at level 1 is being resharded in
parallel to a regular compaction that promotes a sstable at
level 0 (likely to cover full range) to level 1, we will
likely end up with an overlapping at the end.

To fix it, leveled strategy will be aware of compacting sstables
and avoid compaction of sstables when it creates an overlapping
with the ongoing operations (resharding, cleanup, etc).

Reviewed-by: Nadav Har'El <nyh@scylladb.com>
Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>