database: add functions to disable and enable compaction
We have run_with_compaction_disabled, but we want enable_compaction
and disable_compaction for resharding which will select which
sstables to reshard with compaction disabled, or else, we could not
find available sstables that create an overlapping (because they
are already being compacted). We also want to mark as 'compacting'
the sstables we chose to reshard before any compaction run, so
preventing compaction from choosing resharding tables.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>