compaction_strategy: implement resharding strategy for compaction strategies
Strategies other than leveled will reshard one shared sstable at
a time, and the target shard for each job will be chosen in a
round-robin fashion.

For leveled strategy, we will reshard together N adjacent sstables
that belong to same level. N is usually smp::count, but we may
need to add some more sstables due to overlapping.
The reason for that is because resharding one sstable at a time
may result in creation of file for each shard, meaning after
resharding we could end up with NO_SSTABLES*NO_SHARDS.
For level 0 sstables, all unshared sstables will be ignored because
we only use them to deal with overlapping in higher levels.
It chooses the shard for jobs in the same way as other strategies.

These resharding strategies will be used for our new resharding
algorithm.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>