database: implement new sstable resharding algorithm
NOTE: it's not wired yet.

Currently, a shared sstable is rewritten at all shards it belongs
to and only after that, it's deleted. With this new algorithm, a
shared sstable will be read only once and N unshared sstables
will be created, each of them with 1/N of the data. After it's
done, each owner shard will receive its new unshared sstable
replacing its ancestors.

Another benefit is that we'll no longer have resharding resulting
in number of sstables growing considerably after resharding.
A full-sized leveled sstable is usually 160MB, so after resharding,
we could have N files of 160MB/N. Now, leveled strategy will help
resharding. N adjacent sstables of same level will be resharded
together, so we'll end up with N files of N*160MB/N. That is,
we will no longer have the issue described above.

Now for the implementation details:
1) We have to disable compaction while getting resharding jobs
because we need to look at all sstables available and mark them
as compacting before regular compaction picks any of the files
chosen for resharding, or else leveled strategy wouldn't be
able to handle overlapping efficiently.
2) Sstables from different shards are available at shard which
creates resharding jobs by using map reduce. Otherwise, we'd
have to open all files in shard 0, and it's more natural to
just keep it as-is, i.e. each shard open sstables that only
belongs to it.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>