sstables: make sstable deletion safer
When system is booting, sstables that don't belong to current shard
are marked for deletion (which actually means agree on deletion of).
That's okay *now* because all sstables loaded during boot have its
shared field set to true, and a shared sstable is only deleted if
all shards agree on its deletion. A shard which a sstable belongs
to will not agree on its deletion.

An unshared sstable, on the other hand, is deleted if marked for
deletion because it doesn't belong to any other shard.
So if in the future we call set_unshared() for an unshared sstable
before the condition for marking sstable for deletion, unshared
sstables will be incorrectly deleted.
Both sstable::set_unshared() and sstable::is_shared() don't say
a word about it.

To make sstable deletion safer, we'll only proceed to deletion
manager on sstable marked for deletion if it's shared *or* it's
unshared and belongs to current shard.
I also added an assert that get_shards_for_this_sstable() should
never return a empty set, or else, all shards would agree on
deletion of the sstable. That could happen if either sharding
metadata or summary metadata is corrupted.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>