compaction: close all sstable writers for interrupted compaction
Not finishing all writers (even the partial ones) result in undefined
behavior and may lead Scylla to crash in the shutdown.
Interrupted compaction will delete all the new and partial sstables
with compaction::run() catching the exception, calling a function to
do so on created sstables, and rethrowing the exception.

Refs #1900.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>