sstables: call set_unshared for unshared sstables
We weren't calling set_unshared() for loaded unshared sstables, so
sstable::is_shared() would return true for all of them, which is
wrong. I had to make sure that set_unshared() is called only if
sstable belongs to current shard, or else, we could accidently
have the sstable deleted because deletion manager doesn't wait
for all other shards to agree on its deletion.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>