lcs: dont send overlapping sstables to L0
Due to resharding, overlapping may be introduced at level N for a
while, but that will be fixed over time by the strategy itself.
New resharding algorithm will work with sstable from multiple
shards, so it's important that sstables aren't sent to L0, or we
would be basically doing STCS and read performance would be
degraded while most of tables lives in L0.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>