compaction: release space earlier of exhausted input sstables
With LCS, input sstables will not have their space released until
the end of compaction. What we can do is to delete sstables that
are exhausted and which data is all contained in a sealed output
sstable.

LCS will benefit a lot from this change because it usually compacts
1 sstable from level N and 10 tables from level N+1, which means
we'll be able to release 1 table for each 1 table compacted
in average. If compaction fails, nothing bad happens because
if a sstable from level N was removed before exception, there will
be another sstable in level N containing the most recent data
which was merged with data in level N-1.

This will also be helpful for any future strategy that break a run
into non-overlapping fragments.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>