sstables/compaction: propagate sstable replacement to all compaction of same CF
This is needed for parallel compaction to work with sstable run based approach.
That's because regular compaction keeps a cloned set, which may reference the
sstable released. So a compaction A may hold a reference to released sstable
of compaction B, so preventing disk space from being released.

So all replacements are propagated to all compactions of a given column family,
and compactions in turn, including the one which initiated the propagation,
will do the replacement.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>